# apps/server/openapi3.yaml
openapi: 3.0.1
info:
  title: metatile-queue-populator-ui-bff
  description: BFF Service for Metatile Queue Populator UI
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
  /tiles/area:
    post:
      operationId: postTilesByArea
      tags:
        - tiles
      summary: Add all metatiles in given area to queue
      parameters:
        - in: query
          name: force
          required: false
          schema:
            type: boolean
        - in: query
          name: target
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/BboxTilesRequest'
                - $ref: '#/components/schemas/GeometryTilesRequest'
                - $ref: '#/components/schemas/MultiAreaTilesRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
  /tiles/estimate:
    post:
      operationId: estimateTiles
      tags:
        - tiles
      summary: Estimate tile and metatile counts for a given area and zoom range
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - minZoom
                - maxZoom
                - area
              properties:
                minZoom:
                  type: integer
                  minimum: 0
                  maximum: 18
                maxZoom:
                  type: integer
                  minimum: 0
                  maximum: 18
                metatile:
                  type: integer
                  default: 8
                area:
                  oneOf:
                    - type: array
                      items:
                        type: number
                      minItems: 4
                      maxItems: 4
                    - type: object
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - totalMetatiles
                  - totalTiles
                  - metatileSize
                  - breakdown
                properties:
                  totalMetatiles:
                    type: integer
                  totalTiles:
                    type: integer
                  metatileSize:
                    type: integer
                  breakdown:
                    type: array
                    items:
                      type: object
                      properties:
                        zoom:
                          type: integer
                        metatiles:
                          type: integer
                        tiles:
                          type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
  /spatial/convert:
    post:
      operationId: convertSpatial
      tags:
        - spatial
      summary: Convert uploaded Shapefile (.zip), KML, WKT, or GeoJSON to WGS84 GeoJSON
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                format:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - type
                  - features
                properties:
                  type:
                    type: string
                    example: FeatureCollection
                  features:
                    type: array
                    items:
                      type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
  /queue/status:
    get:
      operationId: getQueueStatus
      tags:
        - queue
      summary: Get real-time queue health, active jobs, and status
      parameters:
        - in: query
          name: target
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - timestamp
                  - queues
                  - summary
                properties:
                  status:
                    type: string
                    example: UP
                  timestamp:
                    type: string
                  populatorServiceUrl:
                    type: string
                  targetId:
                    type: string
                  targetName:
                    type: string
                  databaseName:
                    type: string
                  queues:
                    type: array
                    items:
                      type: object
                  summary:
                    type: object
  /queue/metrics:
    get:
      operationId: getQueueMetrics
      tags:
        - queue
      summary: Get queue metrics overview
      parameters:
        - in: query
          name: target
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /history:
    get:
      operationId: getHistory
      tags:
        - history
      summary: Get submission history
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      operationId: recordHistory
      tags:
        - history
      summary: Record new job submission in history
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - parameters
                - status
              properties:
                type:
                  type: string
                  enum: [area, list]
                parameters:
                  type: object
                status:
                  type: string
                  enum: [SUCCESS, FAILED]
                responseMessage:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
  /audit:
    get:
      operationId: getAuditLogs
      tags:
        - audit
      summary: Get system and mutation audit logs
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
        - in: query
          name: method
          required: false
          schema:
            type: string
        - in: query
          name: statusCode
          required: false
          schema:
            type: integer
        - in: query
          name: from
          required: false
          schema:
            type: string
        - in: query
          name: to
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /presets:
    get:
      operationId: getPresets
      tags:
        - presets
      summary: Get list of saved area presets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      operationId: createPreset
      tags:
        - presets
      summary: Save a new area preset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - minZoom
                - maxZoom
                - area
              properties:
                name:
                  type: string
                description:
                  type: string
                minZoom:
                  type: integer
                maxZoom:
                  type: integer
                priority:
                  type: integer
                area:
                  oneOf:
                    - type: array
                      items:
                        type: number
                    - type: object
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
  /presets/{id}:
    delete:
      operationId: deletePreset
      tags:
        - presets
      summary: Delete a saved area preset
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '404':
          $ref: '#/components/responses/NotFound'
  /config/raster:
    get:
      operationId: getRasterConfig
      tags:
        - config
      summary: Get MapColonies Raster CSW and Serving configuration
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - cswUrl
                properties:
                  cswUrl:
                    type: string
                  token:
                    type: string
                  defaultMap:
                    type: object
                    properties:
                      useOsm:
                        type: boolean
                      productId:
                        type: string
                      productType:
                        type: string
  /tiles/list:
    post:
      operationId: postTilesList
      tags:
        - tiles
      summary: Add specific metatiles to queue
      parameters:
        - in: query
          name: force
          required: false
          schema:
            type: boolean
        - in: query
          name: target
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TilesListRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
  /discovery/targets:
    get:
      operationId: getDiscoveryTargets
      tags:
        - discovery
      summary: Retrieve all discovered or configured metatile queue populator targets
      parameters:
        - in: query
          name: refresh
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  targets:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - name
                        - projectName
                        - url
                        - partOf
                        - dbName
                        - source
                        - isDefault
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        projectName:
                          type: string
                        url:
                          type: string
                        partOf:
                          type: string
                        dbName:
                          type: string
                        source:
                          type: string
                          enum:
                            - auto
                            - static
                        isDefault:
                          type: boolean
                        status:
                          type: string
                          enum:
                            - UP
                            - DOWN
                            - UNKNOWN
                        emoji:
                          type: string
        5XX:
          $ref: '#/components/responses/UnexpectedError'
security:
  - {}
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnexpectedError:
      description: Unexpected Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    TilesListRequest:
      type: array
      minItems: 1
      items:
        type: object
        required:
          - z
          - x
          - y
        properties:
          x:
            type: integer
          y:
            type: integer
          z:
            type: integer
          metatile:
            type: integer
            default: 8
    BaseAreaRequest:
      type: object
      required:
        - minZoom
        - maxZoom
      properties:
        minZoom:
          type: integer
          minimum: 0
          maximum: 18
        maxZoom:
          type: integer
          minimum: 0
          maximum: 18
        priority:
          type: integer
          minimum: 0
          default: 0
    BboxTilesRequest:
      allOf:
        - $ref: '#/components/schemas/BaseAreaRequest'
        - type: object
          required:
            - area
          properties:
            area:
              type: array
              items:
                type: number
              minItems: 4
              maxItems: 4
    GeometryTilesRequest:
      allOf:
        - $ref: '#/components/schemas/BaseAreaRequest'
        - type: object
          required:
            - area
          properties:
            area:
              type: object
              required:
                - type
              properties:
                type:
                  type: string
    MultiAreaTilesRequest:
      type: array
      minItems: 1
      items:
        anyOf:
          - $ref: '#/components/schemas/BboxTilesRequest'
          - $ref: '#/components/schemas/GeometryTilesRequest'